requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing & ML
numpy>=1.24.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
PyYAML>=6.0
requests>=2.31.0
tqdm>=4.65.0

# Testing
pytest>=7.0.0

# Add this line to existing requirements
GitPython>=3.1.0 